feat: fs_create_file namespace entry returning a byte-I/O handle
Add HyperDir::fs_create_file, the file counterpart to fs_create_default:
it allocates a UUID, creates the file's hyperfile at layout.file_uri with a
ScatterFirstInterceptor toward the parent, flushes once so the inode object
and the parent's Create scatter both exist, and returns the hyperfile Hyper
handle plus the UUID. This keeps the namespace/scatter logic in hyperdir
while handing the caller a handle for file byte content (read/write/flush).

Also drop an identity cast in the e2e file_nlink helper (st_nlink is already
u64 on the target) to satisfy clippy.

+++ tests/e2e_s3.rs
@@ -177,7 +177,7 @@ async fn create_file(
 
 async fn file_nlink(client: &Client, layout: &HyperDirLayout, bucket: &str, uuid: &Uuid) -> std::io::Result<u64> {
     let st = HyperDir::fs_getattr_fast(client, &layout.file_uri(bucket, uuid)).await?;
-    Ok(st.st_nlink as u64)
+    Ok(st.st_nlink)
 }
 
 async fn file_exists(client: &Client, layout: &HyperDirLayout, bucket: &str, uuid: &Uuid) -> bool {
